+++ App/Model/Waste/BlackBagWaste.php
@@ -4,11 +4,11 @@ namespace App\Model\Waste;
 
 class BlackBagWaste extends AbstractWaste {
 
-  protected int $quantity;
-  protected int $burntCo2;
+  protected float $quantity;
+  protected float $burntCo2;
   protected string $name = "autre";
 
-  public function __construct(int $quantity, int $burntCo2)
+  public function __construct(float $quantity, float $burntCo2)
   {
       $this->quantity = $quantity;
       $this->burntCo2 = $burntCo2;
